Move InputFile into Opal
The component composes InputTypeIn and Button, both Opal; what kept it
app-side was next-intl and a shared click-guard util. The two aria labels
now ride OpalStrings — contract, defaults, bridge, and the catalog keys
move from common.inputFile to opal.inputFile in all nine locales — and
noProp is inlined, since a three-line stopPropagation wrapper is not
worth a cross-package dependency.

Its four consumers import from the barrel.
